Chat thing
You want to download node.js, you want to put the node.js folder,
app.js, index.html, and package,json in a folder. use the command
prompt, navigate to this folder and type 'node app.js' and it should
start the server. Go to a browser and type localhost:4000 (or 3000 I
can't ever remember)